Create String Reduction.cpp
Reduce a string using the following reduction method: Only a, b, and c are allowed in the string and you must take two different adjacent characters and replace with the third (e.g. ab becomes c).  Repeat this method until the string cannot be further reduced and return the length of the final result